# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  internal_handler.criarUsuarioRequest:
    properties:
      login:
        type: string
      nome:
        type: string
      role_id:
        type: string
      senha:
        type: string
    type: object
...
  internal_handler.usuarioResponse:
    properties:
      ativo:
        type: boolean
      id:
        type: string
      login:
        type: string
      nome:
        type: string
      role_id:
        type: string
    type: object
info:
  contact: {}
  description: Sistema de comandas multi-tenant (churrascaria como primeira instância).
    Ver CLAUDE.md e docs/merka-planejamento.md para o contexto completo do domínio.
  title: Merka API
  version: "1.0"
paths:
...
  /usuarios:
    post:
      consumes:
      - application/json
      description: Restrito a Admin Super ou Gestor (permissão "criar_usuario"). Valida
        que role_id pertence ao mesmo tenant. Senha é sempre gravada como hash bcrypt,
        nunca em texto puro.
      parameters:
      - description: Dados do novo usuário
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.criarUsuarioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.usuarioResponse'
        "400":
          description: campo obrigatório ausente ou senha curta
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: token ausente, inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: usuário sem permissão para esta ação
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: perfil (role) não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: já existe um usuário com esse login
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Criar usuário (US-01)
      tags:
      - usuarios
  /usuarios/{id}/desativar:
    patch:
      description: Restrito a Admin Super ou Gestor (permissão "criar_usuario"). Nunca
        deleta — o usuário perde acesso imediatamente (login passa a falhar), mas
        o histórico dele em audit_log permanece intacto.
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: token ausente, inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: usuário sem permissão para esta ação
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: usuário não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Desativar usuário (US-01)
      tags:
      - usuarios
securityDefinitions:
  BearerAuth:
    description: Informe "Bearer <token>" (token obtido em POST /auth/login)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"